Day 5: Using brute force approach for part 2
Using nameref array in part 2,
When reordering the set of pages, I needed the caller to have the corrected version
of the array when it exited.
If I simply pass the array as a string of pages joined by `IFS`, each recursive call would basically have a copy of the array.